+ modified graph construction (kz'02), energy terms (integer weight and thresholding), and datatype for edge capacities from float to int + merge from trunk (3877) to branch

+++ src/maskeditor/segmentation/lazysnapping/maxflow/graph.h
@@ -67,10 +67,11 @@ public:
 	/* Type of edge weights.
 	   Can be changed to char, int, float, double, ... */
 	//typedef short captype;
-    typedef float captype;
+    //typedef float captype;
+	typedef int captype;
 	/* Type of total flow */
-	//typedef int flowtype;
-    typedef float flowtype;
+	typedef int flowtype;
+    //typedef float flowtype;
 
 	typedef void * node_id;
